Unify errorcode
